Haymanot webapi

+++ Application/DTOs/Group/GroupRankingDto.cs
@@ -4,7 +4,7 @@ public class GroupRankingDto : GroupResponseDto
 {
     public int NumberOfProblemsTaken { get; set; }
     public int NumberOfProblemsSolved { get; set; }
-    public int AverageNumberOfProblemsSolved { get; set; }
+    public float AverageNumberOfProblemsSolved { get; set; }
     public double ContestConversionRate { get; set; }
     public int Rank { get; set; }
 }
